fix give_unit test for E3 extract read_factions from read_game, now small and nice.
