+++ scripts/tests/e2/e2features.lua
@@ -12,6 +12,23 @@ function setup()
     eressea.settings.set("rules.peasants.growth.factor", "0")
 end
 
+function test_give_unit()
+  local r = region.create(0, 0, "plain")
+  local f1 = faction.create('elf')
+  local f2 = faction.create('elf')
+  local u1 = unit.create(f1, r)
+  local u2 = unit.create(f2, r)
+  assert_equal(f1, u1.faction)
+  assert_equal(f2, u2.faction)
+  u2.group = 'Experten'
+  assert_equal('Experten', u2.group)
+  u1:add_order("HELFE " .. itoa36(f2.id) .. " GIB")
+  u2:add_order("GIB " .. itoa36(u1.id) .. " EINHEIT")
+  process_orders()
+  assert_equal(f1, u2.faction)
+  assert_nil(u2.group)
+end
+
 function test_study_auto()
     local r = region.create(0, 0, "plain")
     local f = faction.create("human")
